[feature/IndexPage] Added sass support to the esbuild.js build/watch script
For frontend assets, scss modules are bundled into standalone css files in the same directory as any Js file including/requiring them. This will typically occur when using react components that require  sibling scss files.

For backend/server-side files, sass support means that the scss file is resolved to be a css string. This is for the flexibility of having a pure css string that can be handled by the server and served with the text/css, or simply wrapped in a style tag and placed in a React component's jsx.